requirements.txt
Django==5.0.3
viberbot==1.0.12
djangorestframework==3.15.1
requests~=2.31.0
environs~=11.0.0